implement debounce-batch in ssb.db.add
